pgroonga-primary-maintainer: Separate test from test-streaming-replication
Streaming replication is not required.
The test is not changed, just moved.
